backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
requests==2.32.3
beautifulsoup4==4.12.2
python-dotenv==1.0.1
email-validator==2.1.0.post1
